i18n_tool: convert tests to use main instead of Namespace mockup
The logic of tests remains the same but use command line options via
the main() method such as:

        i18n_tool.I18NTool().main([
            '--verbose',
            'translate-desktop',
            '--translations-dir', str(tmpdir),
            '--sources', in_files['source'],
            '--extract-update',

instead of exposing i18n_tool internal details to the tests:

        kwargs = {
            'translations_dir': str(tmpdir),
            'sources': in_files['source'],
            'extract_update': True,
            'compile': False,
            'verbose': logging.DEBUG,
            'version': version.__version__,
        }
        args = argparse.Namespace(**kwargs)
        tool = i18n_tool.I18NTool()
        tool.setup_verbosity(args)

the major drawback of the Namespace mockup is that it must contain all
arguments used by the function, even if the argument has a default
because the arguments are not parsed. It is an unecessary burden
during the test maintenance.